pyproject.toml
[tool.poetry]
name = "my-cafe-python"
version = "0.1.0"
description = ""
authors = ["Sruthi Ganesh <sruthibadal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
django-types = "^0.19.1"
django-stubs-ext = "^5.1.0"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
mysqlclient = "^2.2.4"
setuptools = "^75.1.0"
python-dotenv = "^1.0.1"
django-countries = "^7.6.1"
python-magic = "^0.4.27"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
